fazendo logica para celula ano label text
