+++ Birthdays/Models/BirthdayListModel.swift
@@ -23,20 +23,6 @@ struct BirthdayListModel: Codable {
     }
 }
 
-extension Date {
-    var year: Int {
-        let calendar = Calendar.current
-        let year = calendar.component(.year, from: self)
-        return year
-    }
-
-    var month: Int {
-        let calendar = Calendar.current
-        let month = calendar.component(.month, from: self)
-        return month
-    }
-}
-
 extension Int {
     var asString: String {
         return "\(self)"
